changed rules to the bar version, no more face card rules
After playing slaps with my friends a couple days ago, i realized that when it gets down to the last 2 people the game is almost always a stalemate. This is mainly due to the face cards, which removes a bit of skill from the game and makes it more luck-based. In addition to this, the face card functions were just buggy overall, so I am changing the rules to bar rules, where face cards play like normal cards, except jacks are slappable.